requirements.txt
streamlit
pandas
pyathena
boto3
sqlalchemy
plotly.express
python-dotenv